+++ packages/core/src/service/configService.ts
@@ -33,7 +33,7 @@ export class MidwayConfigService implements IConfigService {
     unittest: 'test',
   };
   private configMergeOrder: Array<ConfigMergeInfo> = [];
-  protected configuration;
+  protected configuration = {};
   protected isReady = false;
   protected externalObject: Record<string, unknown>[] = [];
   protected appInfo: MidwayAppInfo;
@@ -86,10 +86,9 @@ export class MidwayConfigService implements IConfigService {
       } else {
         // object add
         for (const env in dir) {
+          this.getEnvSet(env).add(dir[env]);
           if (this.aliasMap[env]) {
             this.getEnvSet(this.aliasMap[env]).add(dir[env]);
-          } else {
-            this.getEnvSet(env).add(dir[env]);
           }
         }
       }
@@ -187,6 +186,11 @@ export class MidwayConfigService implements IConfigService {
       for (let externalObject of this.externalObject) {
         if (externalObject) {
           externalObject = this.runWithFilter(externalObject);
+          if (!externalObject) {
+            debug('[config]: Filter config and got undefined will be drop it');
+            continue;
+          }
+
           debug('[config]: Loaded external object %j', externalObject);
           extend(true, target, externalObject);
           this.configMergeOrder.push({
@@ -236,7 +240,7 @@ export class MidwayConfigService implements IConfigService {
   }
 
   public clearAllConfig() {
-    this.configuration.clear();
+    this.configuration = {};
   }
 
   public clearConfigMergeOrder() {