+++ packages/oss/src/manager.ts
@@ -8,7 +8,11 @@ import {
 } from '@midwayjs/decorator';
 import * as OSS from 'ali-oss';
 import * as assert from 'assert';
-import { ServiceFactory, delegateTargetPrototypeMethod } from '@midwayjs/core';
+import {
+  ServiceFactory,
+  delegateTargetPrototypeMethod,
+  MidwayCommonError,
+} from '@midwayjs/core';
 import type {
   OSSServiceFactoryReturnType,
   MWOSSClusterOptions,
@@ -72,12 +76,14 @@ export class OSSService implements OSS {
   @Inject()
   private serviceFactory: OSSServiceFactory<OSS>;
 
-  // @ts-expect-error used
   private instance: OSS;
 
   @Init()
   async init() {
     this.instance = this.serviceFactory.get('default');
+    if (!this.instance) {
+      throw new MidwayCommonError('oss default instance not found.');
+    }
   }
 }
 
@@ -99,6 +105,9 @@ export class OSSSTSService implements OSS.STS {
   @Init()
   async init() {
     this.instance = this.serviceFactory.get('default');
+    if (!this.instance) {
+      throw new MidwayCommonError('oss sts default instance not found.');
+    }
   }
 
   async assumeRole(