+++ scripts/rollback/rollback-3.4.13.sh
@@ -1,0 +1,251 @@
+#!/bin/bash
+# timestamp: 1661349037074
+
+npm dist-tag add @midwayjs/fc-starter@3.4.12 latest
+npm dist-tag add @midwayjs/fc-starter@3.4.13 beta
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.4.12 latest
+npm dist-tag add @midwayjs/async-hooks-context-manager@3.4.13 beta
+npm dist-tag add @midwayjs/axios@3.4.12 latest
+npm dist-tag add @midwayjs/axios@3.4.13 beta
+npm dist-tag add @midwayjs/bootstrap@3.4.12 latest
+npm dist-tag add @midwayjs/bootstrap@3.4.13 beta
+npm dist-tag add @midwayjs/cache@3.4.12 latest
+npm dist-tag add @midwayjs/cache@3.4.13 beta
+npm dist-tag add @midwayjs/code-dye@3.4.12 latest
+npm dist-tag add @midwayjs/code-dye@3.4.13 beta
+npm dist-tag add @midwayjs/consul@3.4.12 latest
+npm dist-tag add @midwayjs/consul@3.4.13 beta
+npm dist-tag add @midwayjs/core@3.4.12 latest
+npm dist-tag add @midwayjs/core@3.4.13 beta
+npm dist-tag add @midwayjs/cos@3.4.12 latest
+npm dist-tag add @midwayjs/cos@3.4.13 beta
+npm dist-tag add @midwayjs/cross-domain@3.4.12 latest
+npm dist-tag add @midwayjs/cross-domain@3.4.13 beta
+npm dist-tag add @midwayjs/express-session@3.4.12 latest
+npm dist-tag add @midwayjs/express-session@3.4.13 beta
+npm dist-tag add @midwayjs/faas@3.4.12 latest
+npm dist-tag add @midwayjs/faas@3.4.13 beta
+npm dist-tag add @midwayjs/grpc@3.4.12 latest
+npm dist-tag add @midwayjs/grpc@3.4.13 beta
+npm dist-tag add @midwayjs/http-proxy@3.4.12 latest
+npm dist-tag add @midwayjs/http-proxy@3.4.13 beta
+npm dist-tag add @midwayjs/i18n@3.4.12 latest
+npm dist-tag add @midwayjs/i18n@3.4.13 beta
+npm dist-tag add @midwayjs/info@3.4.12 latest
+npm dist-tag add @midwayjs/info@3.4.13 beta
+npm dist-tag add @midwayjs/jwt@3.4.12 latest
+npm dist-tag add @midwayjs/jwt@3.4.13 beta
+npm dist-tag add @midwayjs/kafka@3.4.12 latest
+npm dist-tag add @midwayjs/kafka@3.4.13 beta
+npm dist-tag add @midwayjs/mikro@3.4.12 latest
+npm dist-tag add @midwayjs/mikro@3.4.13 beta
+npm dist-tag add @midwayjs/mock@3.4.12 latest
+npm dist-tag add @midwayjs/mock@3.4.13 beta
+npm dist-tag add @midwayjs/mongoose@3.4.12 latest
+npm dist-tag add @midwayjs/mongoose@3.4.13 beta
+npm dist-tag add @midwayjs/oss@3.4.12 latest
+npm dist-tag add @midwayjs/oss@3.4.13 beta
+npm dist-tag add @midwayjs/otel@3.4.12 latest
+npm dist-tag add @midwayjs/otel@3.4.13 beta
+npm dist-tag add @midwayjs/passport@3.4.12 latest
+npm dist-tag add @midwayjs/passport@3.4.13 beta
+npm dist-tag add @midwayjs/process-agent@3.4.12 latest
+npm dist-tag add @midwayjs/process-agent@3.4.13 beta
+npm dist-tag add @midwayjs/prometheus-socket-io@3.4.12 latest
+npm dist-tag add @midwayjs/prometheus-socket-io@3.4.13 beta
+npm dist-tag add @midwayjs/prometheus@3.4.12 latest
+npm dist-tag add @midwayjs/prometheus@3.4.13 beta
+npm dist-tag add @midwayjs/rabbitmq@3.4.12 latest
+npm dist-tag add @midwayjs/rabbitmq@3.4.13 beta
+npm dist-tag add @midwayjs/redis@3.4.12 latest
+npm dist-tag add @midwayjs/redis@3.4.13 beta
+npm dist-tag add @midwayjs/security@3.4.12 latest
+npm dist-tag add @midwayjs/security@3.4.13 beta
+npm dist-tag add @midwayjs/sequelize@3.4.12 latest
+npm dist-tag add @midwayjs/sequelize@3.4.13 beta
+npm dist-tag add @midwayjs/session@3.4.12 latest
+npm dist-tag add @midwayjs/session@3.4.13 beta
+npm dist-tag add @midwayjs/socketio@3.4.12 latest
+npm dist-tag add @midwayjs/socketio@3.4.13 beta
+npm dist-tag add @midwayjs/static-file@3.4.12 latest
+npm dist-tag add @midwayjs/static-file@3.4.13 beta
+npm dist-tag add @midwayjs/swagger@3.4.12 latest
+npm dist-tag add @midwayjs/swagger@3.4.13 beta
+npm dist-tag add @midwayjs/tablestore@3.4.12 latest
+npm dist-tag add @midwayjs/tablestore@3.4.13 beta
+npm dist-tag add @midwayjs/task@3.4.12 latest
+npm dist-tag add @midwayjs/task@3.4.13 beta
+npm dist-tag add @midwayjs/typegoose@3.4.12 latest
+npm dist-tag add @midwayjs/typegoose@3.4.13 beta
+npm dist-tag add @midwayjs/typeorm@3.4.12 latest
+npm dist-tag add @midwayjs/typeorm@3.4.13 beta
+npm dist-tag add @midwayjs/upload@3.4.12 latest
+npm dist-tag add @midwayjs/upload@3.4.13 beta
+npm dist-tag add @midwayjs/validate@3.4.12 latest
+npm dist-tag add @midwayjs/validate@3.4.13 beta
+npm dist-tag add @midwayjs/version@3.4.12 latest
+npm dist-tag add @midwayjs/version@3.4.13 beta
+npm dist-tag add @midwayjs/view-ejs@3.4.12 latest
+npm dist-tag add @midwayjs/view-ejs@3.4.13 beta
+npm dist-tag add @midwayjs/view-nunjucks@3.4.12 latest
+npm dist-tag add @midwayjs/view-nunjucks@3.4.13 beta
+npm dist-tag add @midwayjs/view@3.4.12 latest
+npm dist-tag add @midwayjs/view@3.4.13 beta
+npm dist-tag add @midwayjs/express@3.4.12 latest
+npm dist-tag add @midwayjs/express@3.4.13 beta
+npm dist-tag add @midwayjs/koa@3.4.12 latest
+npm dist-tag add @midwayjs/koa@3.4.13 beta
+npm dist-tag add @midwayjs/web@3.4.12 latest
+npm dist-tag add @midwayjs/web@3.4.13 beta
+npm dist-tag add @midwayjs/ws@3.4.12 latest
+npm dist-tag add @midwayjs/ws@3.4.13 beta
+tnpm dist-tag add @midwayjs/fc-starter@3.4.12 latest
+tnpm dist-tag add @midwayjs/fc-starter@3.4.13 beta
+tnpm dist-tag add @midwayjs/async-hooks-context-manager@3.4.12 latest
+tnpm dist-tag add @midwayjs/async-hooks-context-manager@3.4.13 beta
+tnpm dist-tag add @midwayjs/axios@3.4.12 latest
+tnpm dist-tag add @midwayjs/axios@3.4.13 beta
+tnpm dist-tag add @midwayjs/bootstrap@3.4.12 latest
+tnpm dist-tag add @midwayjs/bootstrap@3.4.13 beta
+tnpm dist-tag add @midwayjs/cache@3.4.12 latest
+tnpm dist-tag add @midwayjs/cache@3.4.13 beta
+tnpm dist-tag add @midwayjs/code-dye@3.4.12 latest
+tnpm dist-tag add @midwayjs/code-dye@3.4.13 beta
+tnpm dist-tag add @midwayjs/consul@3.4.12 latest
+tnpm dist-tag add @midwayjs/consul@3.4.13 beta
+tnpm dist-tag add @midwayjs/core@3.4.12 latest
+tnpm dist-tag add @midwayjs/core@3.4.13 beta
+tnpm dist-tag add @midwayjs/cos@3.4.12 latest
+tnpm dist-tag add @midwayjs/cos@3.4.13 beta
+tnpm dist-tag add @midwayjs/cross-domain@3.4.12 latest
+tnpm dist-tag add @midwayjs/cross-domain@3.4.13 beta
+tnpm dist-tag add @midwayjs/express-session@3.4.12 latest
+tnpm dist-tag add @midwayjs/express-session@3.4.13 beta
+tnpm dist-tag add @midwayjs/faas@3.4.12 latest
+tnpm dist-tag add @midwayjs/faas@3.4.13 beta
+tnpm dist-tag add @midwayjs/grpc@3.4.12 latest
+tnpm dist-tag add @midwayjs/grpc@3.4.13 beta
+tnpm dist-tag add @midwayjs/http-proxy@3.4.12 latest
+tnpm dist-tag add @midwayjs/http-proxy@3.4.13 beta
+tnpm dist-tag add @midwayjs/i18n@3.4.12 latest
+tnpm dist-tag add @midwayjs/i18n@3.4.13 beta
+tnpm dist-tag add @midwayjs/info@3.4.12 latest
+tnpm dist-tag add @midwayjs/info@3.4.13 beta
+tnpm dist-tag add @midwayjs/jwt@3.4.12 latest
+tnpm dist-tag add @midwayjs/jwt@3.4.13 beta
+tnpm dist-tag add @midwayjs/kafka@3.4.12 latest
+tnpm dist-tag add @midwayjs/kafka@3.4.13 beta
+tnpm dist-tag add @midwayjs/mikro@3.4.12 latest
+tnpm dist-tag add @midwayjs/mikro@3.4.13 beta
+tnpm dist-tag add @midwayjs/mock@3.4.12 latest
+tnpm dist-tag add @midwayjs/mock@3.4.13 beta
+tnpm dist-tag add @midwayjs/mongoose@3.4.12 latest
+tnpm dist-tag add @midwayjs/mongoose@3.4.13 beta
+tnpm dist-tag add @midwayjs/oss@3.4.12 latest
+tnpm dist-tag add @midwayjs/oss@3.4.13 beta
+tnpm dist-tag add @midwayjs/otel@3.4.12 latest
+tnpm dist-tag add @midwayjs/otel@3.4.13 beta
+tnpm dist-tag add @midwayjs/passport@3.4.12 latest
+tnpm dist-tag add @midwayjs/passport@3.4.13 beta
+tnpm dist-tag add @midwayjs/process-agent@3.4.12 latest
+tnpm dist-tag add @midwayjs/process-agent@3.4.13 beta
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.4.12 latest
+tnpm dist-tag add @midwayjs/prometheus-socket-io@3.4.13 beta
+tnpm dist-tag add @midwayjs/prometheus@3.4.12 latest
+tnpm dist-tag add @midwayjs/prometheus@3.4.13 beta
+tnpm dist-tag add @midwayjs/rabbitmq@3.4.12 latest
+tnpm dist-tag add @midwayjs/rabbitmq@3.4.13 beta
+tnpm dist-tag add @midwayjs/redis@3.4.12 latest
+tnpm dist-tag add @midwayjs/redis@3.4.13 beta
+tnpm dist-tag add @midwayjs/security@3.4.12 latest
+tnpm dist-tag add @midwayjs/security@3.4.13 beta
+tnpm dist-tag add @midwayjs/sequelize@3.4.12 latest
+tnpm dist-tag add @midwayjs/sequelize@3.4.13 beta
+tnpm dist-tag add @midwayjs/session@3.4.12 latest
+tnpm dist-tag add @midwayjs/session@3.4.13 beta
+tnpm dist-tag add @midwayjs/socketio@3.4.12 latest
+tnpm dist-tag add @midwayjs/socketio@3.4.13 beta
+tnpm dist-tag add @midwayjs/static-file@3.4.12 latest
+tnpm dist-tag add @midwayjs/static-file@3.4.13 beta
+tnpm dist-tag add @midwayjs/swagger@3.4.12 latest
+tnpm dist-tag add @midwayjs/swagger@3.4.13 beta
+tnpm dist-tag add @midwayjs/tablestore@3.4.12 latest
+tnpm dist-tag add @midwayjs/tablestore@3.4.13 beta
+tnpm dist-tag add @midwayjs/task@3.4.12 latest
+tnpm dist-tag add @midwayjs/task@3.4.13 beta
+tnpm dist-tag add @midwayjs/typegoose@3.4.12 latest
+tnpm dist-tag add @midwayjs/typegoose@3.4.13 beta
+tnpm dist-tag add @midwayjs/typeorm@3.4.12 latest
+tnpm dist-tag add @midwayjs/typeorm@3.4.13 beta
+tnpm dist-tag add @midwayjs/upload@3.4.12 latest
+tnpm dist-tag add @midwayjs/upload@3.4.13 beta
+tnpm dist-tag add @midwayjs/validate@3.4.12 latest
+tnpm dist-tag add @midwayjs/validate@3.4.13 beta
+tnpm dist-tag add @midwayjs/version@3.4.12 latest
+tnpm dist-tag add @midwayjs/version@3.4.13 beta
+tnpm dist-tag add @midwayjs/view-ejs@3.4.12 latest
+tnpm dist-tag add @midwayjs/view-ejs@3.4.13 beta
+tnpm dist-tag add @midwayjs/view-nunjucks@3.4.12 latest
+tnpm dist-tag add @midwayjs/view-nunjucks@3.4.13 beta
+tnpm dist-tag add @midwayjs/view@3.4.12 latest
+tnpm dist-tag add @midwayjs/view@3.4.13 beta
+tnpm dist-tag add @midwayjs/express@3.4.12 latest
+tnpm dist-tag add @midwayjs/express@3.4.13 beta
+tnpm dist-tag add @midwayjs/koa@3.4.12 latest
+tnpm dist-tag add @midwayjs/koa@3.4.13 beta
+tnpm dist-tag add @midwayjs/web@3.4.12 latest
+tnpm dist-tag add @midwayjs/web@3.4.13 beta
+tnpm dist-tag add @midwayjs/ws@3.4.12 latest
+tnpm dist-tag add @midwayjs/ws@3.4.13 beta
+
+# Changes:
+
+#  - @midwayjs/fc-starter: 3.4.12 => 3.4.13
+#  - @midwayjs/async-hooks-context-manager: 3.4.12 => 3.4.13
+#  - @midwayjs/axios: 3.4.12 => 3.4.13
+#  - @midwayjs/bootstrap: 3.4.12 => 3.4.13
+#  - @midwayjs/cache: 3.4.12 => 3.4.13
+#  - @midwayjs/code-dye: 3.4.12 => 3.4.13
+#  - @midwayjs/consul: 3.4.12 => 3.4.13
+#  - @midwayjs/core: 3.4.12 => 3.4.13
+#  - @midwayjs/cos: 3.4.12 => 3.4.13
+#  - @midwayjs/cross-domain: 3.4.12 => 3.4.13
+#  - @midwayjs/express-session: 3.4.12 => 3.4.13
+#  - @midwayjs/faas: 3.4.12 => 3.4.13
+#  - @midwayjs/grpc: 3.4.12 => 3.4.13
+#  - @midwayjs/http-proxy: 3.4.12 => 3.4.13
+#  - @midwayjs/i18n: 3.4.12 => 3.4.13
+#  - @midwayjs/info: 3.4.12 => 3.4.13
+#  - @midwayjs/jwt: 3.4.12 => 3.4.13
+#  - @midwayjs/kafka: 3.4.12 => 3.4.13
+#  - @midwayjs/mikro: 3.4.12 => 3.4.13
+#  - @midwayjs/mock: 3.4.12 => 3.4.13
+#  - @midwayjs/mongoose: 3.4.12 => 3.4.13
+#  - @midwayjs/oss: 3.4.12 => 3.4.13
+#  - @midwayjs/otel: 3.4.12 => 3.4.13
+#  - @midwayjs/passport: 3.4.12 => 3.4.13
+#  - @midwayjs/process-agent: 3.4.12 => 3.4.13
+#  - @midwayjs/prometheus-socket-io: 3.4.12 => 3.4.13
+#  - @midwayjs/prometheus: 3.4.12 => 3.4.13
+#  - @midwayjs/rabbitmq: 3.4.12 => 3.4.13
+#  - @midwayjs/redis: 3.4.12 => 3.4.13
+#  - @midwayjs/security: 3.4.12 => 3.4.13
+#  - @midwayjs/sequelize: 3.4.12 => 3.4.13
+#  - @midwayjs/session: 3.4.12 => 3.4.13
+#  - @midwayjs/socketio: 3.4.12 => 3.4.13
+#  - @midwayjs/static-file: 3.4.12 => 3.4.13
+#  - @midwayjs/swagger: 3.4.12 => 3.4.13
+#  - @midwayjs/tablestore: 3.4.12 => 3.4.13
+#  - @midwayjs/task: 3.4.12 => 3.4.13
+#  - @midwayjs/typegoose: 3.4.12 => 3.4.13
+#  - @midwayjs/typeorm: 3.4.12 => 3.4.13
+#  - @midwayjs/upload: 3.4.12 => 3.4.13
+#  - @midwayjs/validate: 3.4.12 => 3.4.13
+#  - @midwayjs/version: 3.4.12 => 3.4.13
+#  - @midwayjs/view-ejs: 3.4.12 => 3.4.13
+#  - @midwayjs/view-nunjucks: 3.4.12 => 3.4.13
+#  - @midwayjs/view: 3.4.12 => 3.4.13
+#  - @midwayjs/express: 3.4.12 => 3.4.13
+#  - @midwayjs/koa: 3.4.12 => 3.4.13
+#  - @midwayjs/web: 3.4.12 => 3.4.13
+#  - @midwayjs/ws: 3.4.12 => 3.4.13