+++ CHANGELOG.md
@@ -3,6 +3,45 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.13](https://github.com/midwayjs/midway/compare/v3.4.12...v3.4.13) (2022-08-24)
+
+
+### Bug Fixes
+
+* passport strategy this missing ([#2264](https://github.com/midwayjs/midway/issues/2264)) ([2e5467a](https://github.com/midwayjs/midway/commit/2e5467a7c1cd4b7aa5574ddab624861dea54346b))
+
+
+
+
+
+## [3.4.12](https://github.com/midwayjs/midway/compare/v3.4.11...v3.4.12) (2022-08-20)
+
+
+### Bug Fixes
+
+* grpc context typings ([#2259](https://github.com/midwayjs/midway/issues/2259)) ([9449097](https://github.com/midwayjs/midway/commit/944909705413cebdfe0ae11301e8fd0e536c1edd))
+* service factory client & clients merge ([#2248](https://github.com/midwayjs/midway/issues/2248)) ([cfdee64](https://github.com/midwayjs/midway/commit/cfdee6449cb2770bc238e74fd754b783c331b822))
+
+
+
+
+
+## [3.4.11](https://github.com/midwayjs/midway/compare/v3.4.10...v3.4.11) (2022-08-16)
+
+
+### Bug Fixes
+
+* IMiddleware interface ([#2242](https://github.com/midwayjs/midway/issues/2242)) ([1b435fb](https://github.com/midwayjs/midway/commit/1b435fb86f1ae141f2906b64c236fb8926c4c380))
+
+
+### Features
+
+* **kafka:** update kafka framework & add test demo ([#2236](https://github.com/midwayjs/midway/issues/2236)) ([5eae117](https://github.com/midwayjs/midway/commit/5eae117dc302d1bd56d15564970441e1cd17b185))
+
+
+
+
+
 ## [3.4.10](https://github.com/midwayjs/midway/compare/v3.4.9...v3.4.10) (2022-08-12)